More refactoring of controllers to increase DRYness
